% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-table.R
\name{add_risk}
\alias{add_risk}
\title{Add risk to risk table}
\usage{
add_risk(
  risk_table,
  risk_data,
  cols = NULL,
  scale = NULL,
  join_by = "eu_id",
  overwrite = FALSE
)
}
\arguments{
\item{risk_table}{a risk table initialised with \code{\link[=risk_table]{risk_table()}}}

\item{risk_data}{a data.frame or risk analysis table containing the risk column to
be added to the risk table and the joining column (usually \code{eu_id}).}

\item{risk_col}{}
}
\description{
Function to be used with risk table to add risk columns. The risk table
is designed to store all risks associataed with a riskintro study.
}
\seealso{
Other risk-table: 
\code{\link{remove_risk}()},
\code{\link{risk_table}()}
}
\concept{risk-table}
