% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-entry-point-risk.R
\name{add_source_labels}
\alias{add_source_labels}
\title{Add source labels to display in leaflet}
\usage{
add_source_labels(ep, eps, er)
}
\arguments{
\item{ep}{Entry points table}

\item{eps}{Entry point sources table}

\item{er}{Emission risk table}
}
\description{
Add source labels to display in leaflet
}
