% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-road-access.R
\name{augment_epi_units_from_raster}
\alias{augment_epi_units_from_raster}
\title{Aggregate raster values over polygon areas}
\usage{
augment_epi_units_from_raster(p, r, new_col = "risk")
}
\arguments{
\item{p}{polygon data, usually epidemiologcal units (EU).}

\item{r}{raster data to take the aggregate of each polygon in \code{p}.}

\item{new_col}{name of new column containing aggregated raster data}
}
\value{
returns the polygon data (sf) with a new column containing the
aggregated values from the raster data for each polygon.
}
\description{
Takes the mean of raster values in an area, ignoring any missing values.
}
