% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-road-access.R
\name{augment_epi_units_with_raster}
\alias{augment_epi_units_with_raster}
\title{Aggregate raster values over polygon areas}
\usage{
augment_epi_units_with_raster(
  epi_units,
  raster,
  risk_name = "raster_risk",
  aggregate_fun = c("mean", "max", "min", "sum"),
  reproject = TRUE
)
}
\arguments{
\item{epi_units}{polygon data, usually epidemiologcal units (EU).}

\item{raster}{raster data to take the aggregate of each polygon in \code{p}.}

\item{risk_name}{name of new column containing aggregated raster values}

\item{aggregate_fun}{function to use to aggregate raster values over
epi units area, default is \code{mean}.}

\item{reproject}{whether to reproject \code{raster} to \code{epi_units} projection, default
is TRUE.}
}
\value{
returns the polygon data (sf) with a new column containing the
aggregated values from the raster data for each polygon. The output includes
attributes for use with \code{\link[=plot_risk]{plot_risk()}}: \code{table_name} = "additional_risk",
\code{risk_col} = the value of \code{risk_name}, and \code{scale} = the range of values.
}
\description{
Takes the mean of raster values in an area, ignoring any missing values.
}
