% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-leaflet.R
\name{basemap}
\alias{basemap}
\title{Initilise leaflet object}
\usage{
basemap(...)
}
\arguments{
\item{...}{arguments to pass on to \code{\link[leaflet:leaflet]{leaflet::leaflet()}}}
}
\value{
returns an interactive leaflet visualisation
}
\description{
Initialises leaflet object with Stadia.Outdoors tiles and Esri.WorldImagery, and
controls required to switche between them. Also includes the map scale.

Use \verb{leaflet_add_*()}
functions to add more layers.
}
\examples{
basemap()
}
\seealso{
\code{\link[leaflet:leaflet]{leaflet::leaflet()}}, \code{\link[leaflet:addProviderTiles]{leaflet::addProviderTiles()}},
\code{\link[leaflet:addScaleBar]{leaflet::addScaleBar()}}, \code{\link[leaflet:addLayersControl]{leaflet::addLayersControl()}}
}
