% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-animal-mobility-risk.R
\name{calc_animal_mobility_eu_risk}
\alias{calc_animal_mobility_eu_risk}
\title{Animal mobility risk for each epidemiological unit (EU)}
\usage{
calc_animal_mobility_eu_risk(animal_mobility_points, epi_units, method)
}
\arguments{
\item{animal_mobility_points}{output of \code{calc_animal_mobility_point_risk()}}

\item{epi_units}{study EUs sf object}

\item{method}{the aggregation method to use}
}
\value{
the study EUs with associated introduction risk caluclated from point emission risk
}
\description{
Given the output of the function \code{calc_animal_mobility_point_risk()} (destination
points and their risks) and the study EUs, calculate the risk of introduction into each EU.
}
