% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-animal-mobility-risk.R
\name{calc_animal_mobility_point_risk}
\alias{calc_animal_mobility_point_risk}
\title{Animal mobility destination point risk}
\usage{
calc_animal_mobility_point_risk(animal_mobility, emission_risk, country_iso3)
}
\arguments{
\item{animal_mobility}{the animal mobility dataset}

\item{emission_risk}{emission risk data set}

\item{country_iso3}{the iso3 code for the country in which the study EUs are situated.}
}
\value{
A dataframe of aggregated and weighted risk for each pointin the study EUs.
}
\description{
Calculate the emission risk associated with each animal mobility flow that enters
the epidemiological units (EU). The quantity of each flow associated to a point
inside study EUs is aggregated. The emission risk of a point is then weighted by
the quantity of animals coming from the origin country.
}
