% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-emission-risks.R
\name{calc_emission_risk}
\alias{calc_emission_risk}
\title{Build the weighted emission risk table}
\usage{
calc_emission_risk(emission_risk_factors, weights = get_erf_weights())
}
\arguments{
\item{emission_risk_factors}{A data frame containing risk factor data (default is \code{emission_risk_factors}).}

\item{weights}{A named list of weights corresponding to the columns in \code{dat}. Default weights are applied if not provided.}
}
\value{
A data frame containing:
\itemize{
\item Intermediate scores for each risk domain (\code{sc_epistatus}, \code{sc_survmeasures}, \code{sc_control}, and \code{sc_commerce}).
\item The overall emission risk score (\code{emission_risk}).
}
}
\description{
This function calculates the overall emission risk score (\code{emission_risk}) for a set of countries
based on four domains (each one counts for a certain number of points towards the risk of introduction out of 12 by default) :
\itemize{
\item \strong{Epidemiological status} (3/12):  Time since the last outbreak (\code{sc_epistatus}).
\item \strong{Surveillance measures} (2/12): Effectiveness of implemented surveillance strategies (\code{sc_survmeasures}).
\item \strong{Control measures} (3/12): Effectiveness of disease control measures (\code{sc_control}).
\item \strong{Animal commerce movements} (4/12): Risk from commerce and movement of animals (\code{sc_commerce}).
}

Intermediate scores for each domain are calculated internally:
\itemize{
\item \strong{\code{sc_epistatus}}: The epidemiological status score is based on the time since the last outbreak.
If the disease has not been detected in the last 5 years (\eqn{x > 5}), the score is 0. If the disease is
currently present (\eqn{x = 0}), the score is 3. An exponential decay model smooths the scoring over time:
\deqn{S = 3 \times \exp\left(-x \frac{\log(2)}{5}\right)}
\item \strong{\code{sc_survmeasures}}: Surveillance measures are scored based on the absence of effective measures.
Scores range from 0 (all measures implemented) to 3 (no measures deployed). The default coefficients are:
\itemize{
\item \strong{Active surveillance}: 0.5
\item \strong{Passive surveillance}: 0.5
\item \strong{Risk-based surveillance}: 0.75
\item \strong{Mandatory reporting}: 0.25
}
\item \strong{\code{sc_control}}: Control measures are scored similarly, ranging from 0 (all measures implemented) to 3 (no measures deployed).
Default coefficients are:
\itemize{
\item \strong{Border control}: 1
\item \strong{Culling at outbreak sites}: 0.5
\item \strong{Culling around outbreak sites}: 0.5
\item \strong{Movement zoning and restrictions}: 0.75
\item \strong{Ring vaccination around outbreak sites}: 0.25
}
\item \strong{\code{sc_commerce}}: The risk score for animal commerce movements can take values of 0, 1, or 3, depending on
the risk level identified for trade and movement activities.
}

The overall emission risk score is the weighted sum of these components and is ensured to fall in the range (0, 12].
}
\concept{emission_risk_calculation}
