% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_emission_risk.R
\name{calc_epistatus}
\alias{calc_epistatus}
\title{Calculate epidemiological status risk score (\code{sc_epistatus})}
\usage{
calc_epistatus(dataset, x)
}
\arguments{
\item{dataset}{data.frame like object to add \code{sc_epistatus} column to.}

\item{x}{string of colname, date for last outbreak.}
}
\value{
same as input data with an extra column containing the surveillence
score, where surveillence is in range \deqn{ S \in  (0, 3]}
}
\description{
This function calculates the epidemiological status score (\code{sc_epistatus}) based on the time since the last outbreak.
The formula is:

\deqn{S = 3 \times (1 - \frac{x}{10}) \quad \text{for } x \leq 5}

If the disease has not been detected in the country in the last 5 years (\eqn{x > 5}),
the score is \eqn{S = 0}. If the disease is currently present in the country (\eqn{x = 0}),
the score is \eqn{S = 3}.

To smooth the scoring over the years, an exponential decay model is used:

\deqn{S = 3 \times \exp\left(-x \frac{\log(2)}{5}\right)}
}
\examples{
test_data <- data.frame(
  date_last_outbreak = c(
    Sys.Date(),
    Sys.Date() - 365 * 1,
    Sys.Date() - 365 * 1.5,
    Sys.Date() - 365 * 2,
    Sys.Date() - 365 * 2.5,
    Sys.Date() - 365 * 3,
    Sys.Date() - 365 * 3.5,
    Sys.Date() - 365 * 4,
    Sys.Date() - 365 * 4.5,
    Sys.Date() - 365 * 4.6,
    Sys.Date() - 365 * 4.7,
    Sys.Date() - 365 * 4.8,
    Sys.Date() - 365 * 4.9,
    Sys.Date() - 365 * 4.95,
    Sys.Date() - 365 * 4.99,
    Sys.Date() - 365 * 4.999999999999999999,
    Sys.Date() - 365 * 5,
    Sys.Date() - 365 * 5.5,
    NA # Missing date
  )
)

calc_epistatus(test_data, "date_last_outbreak")
}
\seealso{
Other emission_risk_calculation: 
\code{\link{calc_emission_risk}()}
}
\concept{emission_risk_calculation}
