% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-road-access.R
\name{calc_road_access_risk}
\alias{calc_road_access_risk}
\title{Road access introduction risk analysis}
\usage{
calc_road_access_risk(
  epi_units,
  road_access_raster,
  aggregate_fun = c("mean", "max", "min", "sum")
)
}
\arguments{
\item{epi_units}{epidemiological units dataset}

\item{road_access_raster}{road access raster data contains road accessibility
index. Can be aquired using \code{\link[riskintrodata:download_road_access_raster]{riskintrodata::download_road_access_raster()}} and
\code{\link[terra:rast]{terra::rast()}}.}

\item{aggregate_fun}{function to use to aggregate raster values over
epi units area, default is \code{mean}.}
}
\value{
list with class \code{ri_analysis} that contains:
\enumerate{
\item \code{ri}: an \code{sf} dataset containing the road accessibility risk summarised over each
epidemiological unit. Contains the following columns:
}
\itemize{
\item \code{eu_id}: unique identifier for each epidemiological unit
\item \code{eu_name}: name provided in \code{epi_units} dataset.
\item \code{road_access_risk}: road access risk of inroduction score (unscaled)
\item \code{geometry}: geometry for each epidemiological unit of type \code{POLYGON} or \code{MULTIPOLYGON}.
}

This data can be extracted from the \code{ri_analysis} object with \code{\link[=extract_raster]{extract_raster()}}
\enumerate{
\item \code{raster}: a \code{SpatRaster} containing the cropped raster that covers only
the epidemiological units areas.
}
}
\description{
Calculates the risk of introduction for an animal disease using an
accessibility to roads index.
}
\examples{

\dontrun{
library(riskintroanalysis)
library(riskintrodata)
library(dplyr)
library(terra)
library(sf)
library(ggplot2)
# Example with raw sf files, previously downloaded with geodata::gadm()
tunisia_raw <- read_sf(system.file(
  package = "riskintrodata",
  "samples", "tunisia", "epi_units", "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- apply_mapping(
  tunisia_raw,
  mapping = mapping_epi_units(
    eu_name = "NAME_2",
    geometry = "geom"
  ),
  validate = TRUE
)

road_raster_fp <- download_road_access_raster()
road_raster <- rast(road_raster_fp)

road_access_risk <- calc_road_access_risk(
  epi_units = tunisia,
  road_access_raster = road_raster,
  aggregate_fun = "mean"
)

# plotting the raster data
raster <- extract_raster(road_access_risk)
plot(raster)

# plotting the risk of introduction dataset
ggplot(data=road_access_risk, aes(fill = road_access_risk)) + geom_sf()

}
}
\seealso{
\code{\link[=extract_raster]{extract_raster()}}, \code{\link[riskintrodata:download_road_access_raster]{riskintrodata::download_road_access_raster()}}
}
\concept{default riskintro analysis}
