% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-border-risk.R
\name{calc_weighted_border_risk}
\alias{calc_weighted_border_risk}
\title{Calculate weighted border risk}
\usage{
calc_weighted_border_risk(
  epi_units,
  shared_borders,
  emission_risk,
  ...,
  add_html_lables = FALSE
)
}
\arguments{
\item{epi_units}{epidemiological units dataset}

\item{shared_borders}{shared borders dataset as outputted by \code{\link[=calc_shared_border_lengths]{calc_shared_border_lengths()}}}

\item{emission_risk}{emission risk dataset,}

\item{...}{empty}

\item{add_html_lables}{default FALSE, used for leaflet tooltips in the Shiny app.}
}
\description{
Calculate weighted border risk from required datasets.
}
\details{
Calculates the risk of introduction for each epi unit based on the risk of
emission of neighbouring countries. The score is weighted based on the length of
shared borders.
}
