% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-analysis.R
\name{extract-analysis}
\alias{extract-analysis}
\alias{extract_point_risk}
\alias{extract_flow_risk}
\alias{extract_intro_risk}
\title{Extract analysis subcomponents}
\usage{
extract_point_risk(x)

extract_flow_risk(x)

extract_intro_risk(x)
}
\arguments{
\item{x}{analysis object containing multiple steps/components.}
}
\value{
Returns the analysis step depending on the fuction :
\itemize{
\item \code{extract_point_risk(x)} returns the point risk, usually the first step
in entry point introduction risk analysis. Contains weighted and aggregated
emission risk data for each entry points.
\item \code{extract_flow_risk(x)} returns the flow risk, usually the first step in
animal mobility introduction risk analysis. Contains weighted and aggregated
emission risk data for each flow.
\item \code{extract_intro_risk(x)} returns the dataset containing the risk of introduction
for each epidemiological unit. Usually the final step in all risk analyses.
}
}
\description{
Helper functions to access subcomponents of \code{riskintro_analysis} objets.
}
