% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-table.R
\name{has_risk}
\alias{has_risk}
\title{Check if a risk table has any risks}
\usage{
has_risk(x)
}
\arguments{
\item{x}{a risk table}
}
\value{
TRUE or FALSE
}
\description{
Helper function to check if a risk table has any risks.
}
