% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HTML}
\alias{HTML}
\title{copy of HTML from htmltools}
\usage{
HTML(text, ..., .noWS = NULL)
}
\arguments{
\item{text}{The text value to mark with HTML}

\item{...}{Any additional values to be converted to character and concatenated together}

\item{.noWS}{Character vector used to omit some of the whitespace that would
normally be written around this HTML. Valid options include before, after,
and outside (equivalent to before and end).}
}
\value{
A logical value indicating whether the expression is truthy
}
\description{
A copy of the \code{isTruthy} function from the \code{shiny} package.
}
