% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-entry-points.R
\name{is.boolean}
\alias{is.boolean}
\title{Test whether a vector is boolean}
\usage{
is.boolean(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Either logical vectors or vectors with only 0 and 1 entries count as boolean.
}
\examples{
is.boolean(c(TRUE, TRUE, FALSE))
is.boolean(c(0, 1, 1, 0))
is.boolean(0)
is.boolean(c(0, 1.1, 3))
is.boolean(c("0", "1"))
}
