% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{is.lng}
\alias{is.lng}
\title{Check if a vector is a valid longitude}
\usage{
is.lng(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector of the same length as x
}
\description{
Check if a vector is a valid longitude
}
\keyword{internal}
