% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTruthy}
\alias{isTruthy}
\title{Truthy and falsy values}
\usage{
isTruthy(x)
}
\arguments{
\item{x}{An expression whose truthiness value we want to determine}
}
\value{
A logical value indicating whether the expression is truthy
}
\description{
A copy of the \code{isTruthy} function from the \code{shiny} package.
}
\examples{
isTruthy(1)
isTruthy(character())
isTruthy(character(1))
isTruthy(NULL)
}
