% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-entry-points.R
\name{n_eff_sources}
\alias{n_eff_sources}
\title{Effective number of sources of an entry point}
\usage{
n_eff_sources(x, x_max)
}
\arguments{
\item{x}{Numeric vector. Risk of emission scores of source countries.}

\item{x_max}{Positive number. Maximum possible emission score for a country}
}
\description{
Equivalent number of sources of maximum emisison score.
}
\details{
Each source contributes a quantity between 0 and 1, equal to the ratio between its
emission score and the maximum possible score.
}
\examples{
n_eff_sources(c(12, 12), x_max = 12)  # 2 equivalent sources
n_eff_sources(c(6, 6), x_max = 12)    # 2 half-score sources = 1 equivalent source
}
