% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{apply_mapping}
\alias{mapping_epi_units}
\alias{mapping_entry_points}
\alias{mapping_animal_mobility}
\alias{mapping_emission_risk_factors}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{riskintrodata}{\code{\link[riskintrodata]{apply_mapping}}, \code{\link[riskintrodata]{mapping_animal_mobility}}, \code{\link[riskintrodata]{mapping_emission_risk_factors}}, \code{\link[riskintrodata]{mapping_entry_points}}, \code{\link[riskintrodata]{mapping_epi_units}}}
}}

