% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-table.R
\name{remove_risk}
\alias{remove_risk}
\title{Remove risk from risk table}
\usage{
remove_risk(risk_table, cols)
}
\arguments{
\item{risk_table}{risk table to remove risk from}

\item{cols}{refers to columns (risk scores) in the risk table to remove}
}
\value{
\code{risk_table} with fewer associated risk scores
}
\description{
Remove risk from risk table
}
\examples{
library(riskintrodata)
library(riskintroanalysis)

# Get epi_units dataset
tunisia_raw <- sf::read_sf(system.file(
  package = "riskintrodata",
  "samples",
  "tunisia",
  "epi_units",
  "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- validate_dataset(
  x = tunisia_raw,
  table_name = "epi_units",
  eu_name = "NAME_2",
  geometry = "geom"
) |>
  extract_dataset()

# Create risk table
rt <- risk_table(tunisia, scale = c(0, 100))

# Create risk
entry_points_fp <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "entry_points",
    "BORDER_CROSSING_POINTS.csv"
  )

entry_points <- readr::read_csv(entry_points_fp)
entry_points <- validate_dataset(
  x = entry_points,
  table_name = "entry_points",
  point_name = "NAME",
  lng = "LONGITUDE_X",
  lat = "LATITUDE_Y",
  mode = "MODE",
  type = "TYPE",
  sources = "SOURCES"
) |>
  extract_dataset()

ri_entry_points <- calc_entry_point_risk(
  entry_points = entry_points,
  epi_units = tunisia,
  emission_risk = calc_emission_risk(
    emission_risk_factors = get_wahis_erf(
      disease = "Anthrax",
      species = "Cattle",
      animal_category = "Domestic"
    )
  )
) |>
  rescale_risk_scores()

rt <- add_risk(risk_table = rt, risk_data = ri_entry_points)
rt

rt <- remove_risk(rt, cols = "entry_points_risk")
rt
}
\seealso{
Other risk-table: 
\code{\link{add_risk}()},
\code{\link{risk_table}()}
}
\concept{risk-table}
