% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-rescale_risk.R
\name{rescale_risk_scores}
\alias{rescale_risk_scores}
\title{Scaling range transformation function (re-scaling)}
\usage{
rescale_risk_scores(
  data,
  cols,
  ...,
  from,
  to = c(0, 100),
  method = c("linear", "quadratic", "exponential", "sigmoid"),
  inverse = FALSE,
  names_prefix = NULL,
  names_to = NULL,
  keep_cols = TRUE
)
}
\arguments{
\item{data}{dataset to add scaled risk column to.}

\item{cols}{Column containing numeric vector of risk values in the range [0, 12] OR a named
vector, names will be used as the \code{names_to} arg, giving new names to rescaled columns.}

\item{...}{keep empty}

\item{from}{existing range of possible values for \code{risk_col} that will be converted to}

\item{to}{new range of possible values for \code{risk_col}, by default it is 0 to 100.}

\item{method}{The scaling method to apply. Options are \code{"linear"}, \code{"quadratic"},
\code{"exponential"}, \code{"sigmoid"}, or \code{"complementary"}.}

\item{inverse}{boolean to inverse the risk values, i.e. low becomes high, and
high becomes low. Similar to the \code{"complementary"} value of method, but can be
added to quadratic.}

\item{names_prefix}{string, prefix \code{cols} names to have new names for scaled columns.}

\item{names_to}{string vector, provide new name for rescaled columns.}

\item{keep_cols}{default TRUE, whether to keep \code{cols} columns after rescaling is done.}
}
\value{
A dataset with new column containing  numeric vector of scaled values in the range [0, 100].
}
\description{
This function supports multiple scaling methods to transform risk values from
an existing range to  another. By default the output range is between 0 and 100.
}
\details{
Below are the equations used for each method:

\strong{Linear Scaling}
\deqn{f(x) = \frac{x - 0}{12 - 0} \cdot (100 - 0) + 0}

\strong{Quadratic Scaling}
\deqn{f(x) = \left(\frac{x - 0}{12 - 0}\right)^2 \cdot (100 - 0) + 0}

\strong{Exponential Scaling}
\deqn{f(x) = \frac{\exp(x / 12) - 1}{\exp(1) - 1} \cdot (100 - 0) + 0}

\strong{Sigmoid Scaling}
\deqn{f(x) = \frac{100}{1 + \exp(-10 \cdot (\frac{x}{12} - 0.5))}}
}
\examples{
# rescale_risk -----
library(riskintrodata)
library(riskintroanalysis)

dat <- data.frame(RISK = seq(0, 12, by = 0.01))

dat_scaled <- rescale_risk_scores(
  data = dat,
  cols = "RISK",
  method = "linear",
  from = c(0, 12),
  to = c(0, 100),
  names_prefix = "sigmoid_"
)

ggplot(dat_scaled, aes(x = RISK, y  = sigmoid_RISK)) +
  geom_point()
}
