% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk_interactive.R
\name{riskintro-leaflet-legends}
\alias{riskintro-leaflet-legends}
\alias{riskPalette}
\alias{addRiskLegend}
\alias{scorePalette}
\alias{addScoreLegend}
\title{Legends and palettes for RiskIntro leaflets}
\usage{
riskPalette(scale = c(0, 100))

addRiskLegend(ll, scale, title, opacity = 0.7, layerId = "leaflet_legend")

scorePalette(scale = c(0, 100))

addScoreLegend(
  ll,
  scale = c(0, 100),
  title = "Score",
  opacity = 0.7,
  layerId = "leaflet_legend"
)
}
\arguments{
\item{scale}{scale used for domain.}

\item{ll}{leaflet object or proxy to add legend to.}

\item{title}{the legend title}

\item{opacity}{the opacity of colors}

\item{layerId}{the ID of the legend; subsequent calls to \code{addLegend()}
or \code{addControl()} with the same \code{layerId} will replace this
legend. The ID can also be used with \code{removeControl()}.}
}
\description{
Interactive plots use these pallettes and legends. Risks use the viridis
palette while scores use the inferno palette. These functions are built on
\code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}} and \code{\link[leaflet:addLegend]{leaflet::addLegend()}}.
}
