% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-entry-point-risk.R
\name{scale_controlled}
\alias{scale_controlled}
\title{Convert controlled to equivalent non-controlled entry point exposure}
\usage{
scale_controlled(x, alpha = 1, beta = 1, lambda = 3)
}
\arguments{
\item{x}{numeric vector of controlled entry points}

\item{alpha}{(coef_legal) Number > 0. Scaling factor of legal sources in the latent scale.}

\item{beta}{(coef_illegal) Number > 0. Scaling factor of illegal sources in the latent scale.}

\item{lambda}{(illegal_factor) Number > 1. Relative risk of an illegal entry point with respect
to a legal one.}
}
\description{
Used for univariate scaling of the entry points dataset.
}
