% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-entry-point-risk.R
\name{scale_entry_points}
\alias{scale_entry_points}
\title{Scale risk of entry points}
\usage{
scale_entry_points(
  x_legal,
  x_illegal,
  illegal_factor,
  coef_legal,
  coef_illegal,
  max_risk
)
}
\arguments{
\item{x_legal}{Numeric vector. Effective numbers of sources for each legal entry point.}

\item{x_illegal}{Numeric vector. Effective numbers of sources for each illegal entry point.}

\item{illegal_factor}{Number > 1. Relative risk of an illegal entry point with respect
to a legal one.}

\item{coef_legal}{Number > 1. Scaling factor of legal sources in the latent scale.}

\item{coef_illegal}{Number > 1. Scaling factor of illegal sources in the latent scale.}

\item{max_risk}{Number > 1. Maximum assymptotically atteinable risk.}
}
\description{
Scale the effective numbers of legal and illegal sources into the risk scale.
}
\examples{
library(ggplot2)
library(tidyr)
case_dat <- tidyr::expand_grid(
  xl = 0:5,
  xi = 0:5
)
case_dat |>
  transform(
    risk = scale_entry_points(xl, xi, 3, 1, 1, 100)
  ) |>
  ggplot(aes(xl, xi, fill = risk)) +
  geom_raster() +
  coord_fixed() +
  scale_fill_viridis_c(option = "D") +
  labs(
    x = "N legal eps",
    y = "N illegal eps",
    fill = "Risk"
  )
}
