% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale-risk-scores.R, R/utils-rescaling.R
\name{sigmoid}
\alias{sigmoid}
\alias{inv_sigmoid}
\title{Sigmoid and inverse sigmoid}
\usage{
sigmoid(x)

inv_sigmoid(x)

inv_sigmoid(x)
}
\arguments{
\item{x}{Numeric vector with values between -1 and 1.}
}
\value{
A numeric vector with values between -1 and 1.

A numeric vector with real values.
}
\description{
We use the logistic function, rescaled to map non-negative values into [0, 1).

The invesrse of the sigmoid function.
}
\details{
The function increases asymptotically up to 1, taking the values .46, .76, .90 and .96
at 1, 2, 3 and 4 respectively.
}
\examples{
x <- (0:500)/100
plot(x, sigmoid(x), type = 'l')
x <- c(0.01, 0.88, 0.247)
inv_sigmoid(x)
}
\keyword{internal}
