% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-summary.R
\name{summarise_risk_scores}
\alias{summarise_risk_scores}
\title{Summarise risk score}
\usage{
summarise_risk_scores(
  risk_table,
  cols,
  method = c("mean", "max", "min", "median")
)
}
\arguments{
\item{risk_table}{dataset containing all risk data for each epidemiological unit}

\item{cols}{columns to summarise}

\item{method}{summary method such as \code{"mean"} or \code{"max"}.}
}
\description{
Summarise risk scores across columns using the chosen method.
}
