% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-road-access.R
\name{updateRoadAccessLeaflet}
\alias{updateRoadAccessLeaflet}
\title{Plot risk access}
\usage{
updateRoadAccessLeaflet(ll, dat, r)
}
\arguments{
\item{ll}{leaflet or leaflet proxy to add polygons and raster to}

\item{dat}{sf dataset with risks}

\item{r}{splatRaster obj}
}
\value{
leaflet plot with legend
}
\description{
Plot risk access
}
