% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-road-access.R
\name{augment_epi_units_from_raster}
\alias{augment_epi_units_from_raster}
\title{Aggregate raster values over polygon areas}
\usage{
augment_epi_units_from_raster(
  epi_units,
  raster,
  risk_name = "raster_risk",
  aggregate_fun = c("mean", "max", "min", "sum")
)
}
\arguments{
\item{epi_units}{polygon data, usually epidemiologcal units (EU).}

\item{raster}{raster data to take the aggregate of each polygon in \code{p}.}

\item{risk_name}{name of new column containing aggregated raster values}

\item{aggregate_fun}{function to use to aggregate raster values over
epi units area, default is \code{mean}.}
}
\value{
returns the polygon data (sf) with a new column containing the
aggregated values from the raster data for each polygon.
}
\description{
Takes the mean of raster values in an area, ignoring any missing values.
}
