% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-animal-mobility-risk.R
\name{calc_animal_mobility_risk}
\alias{calc_animal_mobility_risk}
\title{Calculate animal mobility risk}
\usage{
calc_animal_mobility_risk(animal_mobility, emission_risk, epi_units, method)
}
\arguments{
\item{animal_mobility}{animal moblity dataset as formatted and validated by
\code{\link[=apply_mapping]{apply_mapping()}} and \code{\link[=mapping_animal_mobility]{mapping_animal_mobility()}}}

\item{emission_risk}{emission risk dataset from \link{calc_emission_risk}}

\item{epi_units}{epidemiological units dataset}

\item{method}{aggregation method for eu risk}
}
\description{
Calculate the risk of introduction associated with legal international animal
commerce entering epidemiological units.

The animal mobility dataset contains a single legal animal commerce route per
row. Each row has the starting points and arrival point. Emission risk is
weighted by quantity if provided.

Once each arrival point has an averaged emission risk based on the emission score
from each source country the \code{method} argument is used to aggregate that further,
giving a risk of introduction for each epidemiological unit.
}
