% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-border-risk.R
\name{calc_border_lengths}
\alias{calc_border_lengths}
\title{Get Shared Borders Between EUs and Bordering Countries}
\usage{
calc_border_lengths(epi_units, eu_country_iso3 = NULL, neighbours = NULL)
}
\arguments{
\item{epi_units}{An \code{sf} object containing the epidemiological units dataset.}

\item{eu_country_iso3}{A string specifying the iso3 code of the epidemiological units
country. \code{NULL} by default, this argument is found by locating \code{epi_units} on a
world map. Only required when \code{epi_units} does not correspond to an actual country.}

\item{neighbours}{A string vector of iso3 codes corresponding to neighbouring
countries of the \code{epi_units}. \code{NULL} by default, neighbours are found using
the \link[riskintrodata:neighbours_table]{riskintrodata::neighbours_table}. If values are provide for this parameter,
then any argument for \code{eu_country_iso3} is not used.}
}
\value{
An \code{sf} object containing shared borders between EUs and bordering countries,
with calculated border lengths and weights. The table is containing the following columns:
\itemize{
\item \code{eu_id}: ID of the epidemiological unit.
\item \code{bc_id}: ISO3 of the bordering country.
\item \code{border_length}: Length of the shared border, use \link[units:units]{units::units} to get
the units. It is expected in km.
\item \code{geometry}: The geometry column containing the shared border, a \code{MULTILINESTRING}
or \code{LINESTRING}.
\item \code{weight}: The risk weighting coefficient based on the length of the border.
}
}
\description{
This function identifies shared borders between epidemiological units (EUs) and neighbouring countries.
It accounts for small discrepancies in border alignments and adjusts accordingly using an
algorithm that loops over all overlapping and divergent polygons.

This can take some time to run.

Distances are calculated using projection crs 6933 and uses s2. However, output remains WGS84.
}
\details{
This is the first step in the border risk analysis method, the outputs of this
function should be passed on to \code{\link[=calc_border_risk]{calc_border_risk()}}.
}
\examples{
#
library(sf)
library(dplyr)
library(riskintrodata)
library(riskintroanalysis)

tunisia_raw <- sf::read_sf(system.file(
  package = "riskintrodata",
  "samples",
  "tunisia",
  "epi_units",
  "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- validate_dataset_content(
  x = tunisia_raw,
  table_name = "epi_units",
  eu_name = "NAME_2",
  geometry = "geom"
) |>
  extract_dataset()

# Run function to get shared borders
shared_borders <- calc_border_lengths(
  epi_units = tunisia
)

wahis_factors <- riskintrodata::get_wahis_erf(
  disease = "Anthrax",
  species = "Cattle",
  animal_category = "Domestic"
)

emission_risk_table <- calc_emission_risk(wahis_factors)

ri_border <- calc_border_risk(
  epi_units = tunisia,
  shared_borders = shared_borders,
  emission_risk = emission_risk_table
)
ri_border

border_lines <- extract_border(ri_border)
border_lines
}
