% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-border-risk.R
\name{calc_border_risk}
\alias{calc_border_risk}
\title{Calculate weighted border risk}
\usage{
calc_border_risk(epi_units, shared_borders, emission_risk)
}
\arguments{
\item{epi_units}{epidemiological units dataset}

\item{shared_borders}{shared borders dataset as outputted by \code{\link[=calc_border_lengths]{calc_border_lengths()}}}

\item{emission_risk}{emission risk dataset,}
}
\value{
An \code{sf} object containing the border risk score associated with each
epidemiological unit. The table is containing the following columns:
\itemize{
\item \code{eu_id}: ID of the epidemiological unit. This will be used as a join key to the
\code{epi_units} table and \code{risk_table}.
\item \code{eu_name}: Name of the epidemiological unit. This comes from the \code{epi_units} table.
\item \code{borders}: Total length of shared borders with neighbouring countries in km.
\item \code{border_risk}: The risk of introduction based on the emission risk of neighbouring countries,
its values are in the range [0, 12].
\item \code{geometry}: The geometry column containing the shared border, a \code{MULTIPOLYGON}.
}

This dataset also has a \strong{number of attributes}, the first is a dataset, the other 3
are used in other functions from \code{riskintroanalysis} to make passing dataset metadata
between functions more user-friendly.
\enumerate{
\item \code{borders}: an \code{sf} object containing shared borders between EUs and bordering countries.
The table is containing the following columns:
\itemize{
\item \code{eu_id}: ID of the epidemiological unit.
\item \code{bc_id}: ISO3 of the bordering country.
\item \code{border_length}: Length of the shared border, use \link[units:units]{units::units} to get
the units. It is expected in km.
\item \code{geometry}: The geometry column containing the shared border, a \code{MULTILINESTRING} or a
\code{LINESTRING}.
}
\item \code{table_name = "border_risk"}: the name of the table in the riskintroanalysis
data model. This is used by other functions in  \code{riskintroanalysis}, such as \code{plot_risk}.
\item \code{risk_col = "border_risk"}: the name of the risk score column in the dataset.
This is used by other functions in  \code{riskintroanalysis}, such as \code{plot_risk}.
\item \code{scale = c(0,12)}: the scale of \code{risk_col} used for other functions such
as \code{rescale_risk_score} and \code{plot_risk}.
}
}
\description{
Calculates the risk of introduction for each epidemiological unit based on the risk of
emission of neighbouring countries. The score is weighted based on the length of
shared borders.
}
\examples{
#
library(sf)
library(dplyr)
library(riskintrodata)
library(riskintroanalysis)

tunisia_raw <- sf::read_sf(system.file(
  package = "riskintrodata",
  "samples",
  "tunisia",
  "epi_units",
  "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- validate_dataset(
  x = tunisia_raw,
  table_name = "epi_units",
  eu_name = "NAME_2",
  geometry = "geom"
) |>
  extract_dataset()

# Run function to get shared borders
shared_borders <- calc_border_lengths(
  epi_units = tunisia
)


algeria <- erf_row(
  iso3 = "DZA",
  country = "Algeria",
  disease = "Avian infectious laryngotracheitis",
  animal_category = "Domestic",
  species = "Birds",
  disease_notification = 0,
  targeted_surveillance = 1,
  general_surveillance = 0,
  screening = 1,
  precautions_at_the_borders = 1,
  slaughter = 1,
  selective_killing_and_disposal = 1,
  zoning = 1,
  official_vaccination = 1,
  last_outbreak_end_date = as.Date("30/06/2023"),
  commerce_illegal = 0L,
  commerce_legal = 0L
)

libya <- erf_row(
  iso3 = "LBY",
  country = "Libya",
  disease = "Avian infectious laryngotracheitis",
  animal_category = "Domestic",
  species = "Birds",
  disease_notification = TRUE,
  targeted_surveillance = 1,
  general_surveillance = 0,
  screening = 1,
  precautions_at_the_borders = 0,
  slaughter = 1,
  selective_killing_and_disposal = 1,
  zoning = 1,
  official_vaccination = 1,
  last_outbreak_end_date = as.Date("30/06/2019"),
  commerce_illegal = 0L,
  commerce_legal = 1
)

wahis_erf <- get_wahis_erf(
  disease = "Avian infectious laryngotracheitis",
  animal_category = "Domestic",
  species = "Birds"
)

emission_risk_factors <- dplyr::bind_rows(
  algeria,
  libya,
  wahis_erf
)
emission_risk_table <- calc_emission_risk(emission_risk_factors)

ri_border <- calc_border_risk(
  epi_units = tunisia,
  shared_borders = shared_borders,
  emission_risk = emission_risk_table
)
ri_border

border_lines <- extract_border(ri_border)
border_lines

plot_risk_interactive(ri_border)
}
