% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_emission_risk.R
\name{calc_emission_risk}
\alias{calc_emission_risk}
\title{Build the weighted emission risk table}
\usage{
calc_emission_risk(
  emission_risk_factors,
  weights = get_erf_weights(),
  keep_scores = TRUE
)
}
\arguments{
\item{emission_risk_factors}{A data frame containing risk factor data. Generally,
this data will come from \code{\link[riskintrodata:get_wahis_erf]{riskintrodata::get_wahis_erf()}}. The dataset should be validated and
have \code{table_name} attribute equal to \code{"emission_risk_factors"}.}

\item{weights}{A named list of weights corresponding to the following columns in
\code{emission_risk_factors} (and their default weights):
\itemize{
\item \code{disease_notification} (0.25)
\item \code{targeted_surveillance} (0.5)
\item \code{general_surveillance} (0.5)
\item \code{screening} (0.75)
\item \code{precautions_at_the_borders} (1)
\item \code{slaughter} (0.5)
\item \code{selective_killing_and_disposal} (0.5)
\item \code{zoning} (0.75)
\item \code{official_vaccination} (0.25)
}

The sum of the weights should add up to exactly 5, as these factors correspond
to the weights contributing to \code{sc_survmeasures} and \code{sc_control}.}

\item{keep_scores}{whether to keep or drop \verb{sc_*} columns, \code{emission_risk} column
is always kept.}
}
\value{
A tibble containing the following columns:
\itemize{
\item \code{iso3} identifis the country
\item \code{country} country name (from emission risk factors dataset)
\item \code{disease} disease being studied (from emission risk factors dataset)
\item \code{animal_category } animal_category  being studied (from emission risk factors dataset)
\item \code{species} species being studied (from emission risk factors dataset)
\item \code{data_source} data source for emission risk factors (from emission risk factors dataset)
\item \code{sc_survmeasures} as detailed above.
\item \code{sc_control} as detailed above.
\item \code{sc_commerce} as detailed above.
\item \code{sc_epistatus} as detailed above.
\item \code{emission_risk} as detailed above.
}

This dataset also has a \strong{number of attributes} that are used in other
functions from \code{riskintroanalysis} to make passing dataset metadata between
functions more user-friendly. Used mainly in used by \code{\link[=plot_risk]{plot_risk()}} and
\code{\link[=rescale_risk_scores]{rescale_risk_scores()}}.
\itemize{
\item \code{table_name = "emission_risk_scores"}
\item \code{risk_col = "emission_risk"}
\item \code{scale = c(0, 12)}
\item \code{table_validated = TRUE}
}
}
\description{
This function calculates the overall emission risk score (\code{emission_risk}) for a set of countries
based on four domains (each one counts for a certain number of points towards the risk of introduction out of 12 by default) :
\itemize{
\item \strong{Epidemiological status} (3/12):  Time since the last outbreak (\code{sc_epistatus}).
\item \strong{Surveillance measures} (2/12): Effectiveness of implemented surveillance strategies (\code{sc_survmeasures}).
\item \strong{Control measures} (3/12): Effectiveness of disease control measures (\code{sc_control}).
\item \strong{Animal commerce movements} (4/12): Risk from commerce and movement of animals (\code{sc_commerce}).
}

The emission risk is the sum of each of the above.

Each of the scores is calculated from the emission risk factors and some are weighted
based on the emission risk factor weights (\code{weights} parameter).
\itemize{
\item \strong{\code{sc_epistatus}}: The epidemiological status score is based on the time since the last outbreak
and accounds for 3 out of 12 of the final emission risk score.
If the disease has not been detected in the last 5 years (\eqn{x > 5}), the score is 0. If the disease is
currently present (\eqn{x = 0}), the score is 3. An exponential decay model smooths the scoring over time:
\deqn{S = 3 \times \exp\left(-x \frac{\log(2)}{5}\right)}
\item \strong{\code{sc_survmeasures}}: Surveillance measures are scored based on the absence
of effective measures and accounts for 2 out of 12 og the final emission risk
score. The following risk factors contribute
to this score:
\itemize{
\item Active surveillance
\item Passive surveillance
\item Risk-based surveillance
\item Mandatory reporting
}
\item \strong{\code{sc_control}}: Control measures are scored similarly, and account for
3 out of 12 of the emission risk score. The following risk factors contribute
to this score:
\itemize{
\item Border control
\item Culling at outbreak sites
\item Culling around outbreak sites
\item Movement zoning and restrictions
\item Ring vaccination around outbreak sites
}
\item \strong{\code{sc_commerce}}: The risk score for animal commerce movements can take
values of 0, 1, 3, or 4. It is the sum of the following:
\itemize{
\item 0 means there is no legal or illegal trade
\item legal trade present adds 1 to this score
\item illegal trade present adds 3 to this score
}
}

The overall emission risk score is the sum of each weighted risk factor,
the final emission risk score is in the range of (0, 12].
}
\section{NA values}{


NA values are considered to have the highest level of risk. For example if
\code{disease_notification} for a country in the emission risk factors dataset is
\code{NA} then disease notification will be considered as an unimplemented policy
for the purpose of calculating emission risk scores.
}

\examples{
library(riskintrodata)
library(riskintroanalysis)

wahis_erf <- get_wahis_erf(
  disease = "Avian infectious laryngotracheitis",
  animal_category = "Domestic",
  species = "Birds"
)

emission_risk_table <- calc_emission_risk(
  emission_risk_factors = wahis_erf
)
}
\concept{emission_risk_calculation}
