% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-entry-point-risk.R
\name{calc_entry_point_risk}
\alias{calc_entry_point_risk}
\title{Calculate entry point risk}
\usage{
calc_entry_point_risk(entry_points, epi_units, emission_risk)
}
\arguments{
\item{entry_points}{The entry points dataset as formatted and validated by \code{\link[=apply_mapping]{apply_mapping()}} and
\code{\link[=mapping_entry_points]{mapping_entry_points()}}. This should be an \code{sf} object containing points and emission risks.}

\item{epi_units}{The epidemiological units dataset as formatted and validated by \code{\link[=apply_mapping]{apply_mapping()}} and
\code{\link[=mapping_epi_units]{mapping_epi_units()}}. This should be an \code{sf} object with polygons.}

\item{emission_risk}{The emission risk dataset as returned by the \code{\link[=calc_emission_risk]{calc_emission_risk()}}
function.}
}
\value{
returns a list of class \code{ri_analysis} containing two datasets:
\enumerate{
\item Points (name: \code{points}) dataset contains the aggregated emission risk score for each
entry point.
\item Epidemiological units (name: \code{epi_units}) dataset containing the weighted
average of each of points' emission risk scores within its area, this is the
risk of introduction by entry points.
}
}
\description{
Calculating the entry point risk associated with each epidemiological unit.

The first step is calculating the points dataset: each
point located within an epidemiological unit is linked to one or more source countries from
which animals enter into the epidemiological unit. For each point the emission
risk of the source countries is averaged. This provides an average emission risk score
for each point.

The second step is creating the epidemiologcal units introduction risk dataset by
joining th points dataset with epidemiological units and
finding the average emission risk score for each entry points within each
area. This gives a risk of introduction for each epidemiological unit.
}
