% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-emission-risks.R
\name{get_erf_weights}
\alias{get_erf_weights}
\title{Get Default Emission Risk Weights}
\usage{
get_erf_weights()
}
\value{
named list of emission risk weights \link[riskintrodata:emission_risk_weights]{riskintrodata::emission_risk_weights}
}
\description{
Get Default Emission Risk Weights
}
