% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk.R, R/plot_risk_interactive.R
\name{plot_risk}
\alias{plot_risk}
\alias{plot_entry_points}
\alias{plot_animal_mobility}
\alias{plot_road_access}
\alias{plot_border_risk}
\alias{plot_epi_units}
\alias{plot_risk_table}
\alias{plot_risk_interactive}
\alias{plot_entry_points_interactive}
\alias{plot_animal_mobility_interactive}
\alias{plot_road_access_interactive}
\alias{plot_border_risk_interactive}
\alias{plot_epi_units_interactive}
\alias{plot_risk_table_interactive}
\title{Plot risk}
\usage{
plot_risk(
  dataset,
  risk_name = NULL,
  risk_col = NULL,
  scale = NULL,
  interactive = FALSE
)

plot_entry_points(dataset, scale, risk_col)

plot_animal_mobility(dataset, scale, risk_col)

plot_road_access(dataset, scale, risk_col)

plot_border_risk(dataset, scale, risk_col)

plot_epi_units(dataset, scale, risk_col)

plot_risk_table(dataset, scale, risk_col)

plot_risk_interactive(
  dataset,
  risk_name = NULL,
  risk_col = NULL,
  scale = NULL,
  ll = basemap()
)

plot_entry_points_interactive(dataset, scale, risk_col, ll = basemap())

plot_animal_mobility_interactive(dataset, scale, risk_col, ll = basemap())

plot_road_access_interactive(dataset, scale, risk_col, ll = basemap())

plot_border_risk_interactive(dataset, scale, risk_col, ll = basemap())

plot_epi_units_interactive(dataset, scale, risk_col, ll = basemap())

plot_risk_table_interactive(dataset, scale, risk_col, ll = basemap())
}
\arguments{
\item{dataset}{sf object with risk table data}

\item{risk_name}{risk_name used to determine which plotting function to use,
this is NULL by default and is inferred from \code{attr(x, "table_name")}.}

\item{risk_col}{character string naming the risk column to visualize}

\item{scale}{numeric vector of length 2 defining the risk scale}

\item{interactive}{default to FALSE for ggplot2 visuals, TRUE returns
interactive leaflet.}

\item{ll}{leaflet object to add layers to}
}
\value{
ggplot2 object with the appropriate visualisation if interactive is FALSE,
otherwise, leaflet object

leaflet map object

leaflet map object

leaflet map object

leaflet map object

leaflet map object

leaflet map object
}
\description{
Helper function to plot riskintro's default risks using ggplot2.
}
\examples{
tunisia_raw <- sf::read_sf(system.file(
  package = "riskintrodata",
  "samples",
  "tunisia",
  "epi_units",
  "tunisia_adm2_raw.gpkg"
))

# Apply mapping to prepare colnames and validate dataset
tunisia <- validate_dataset(
  x = tunisia_raw,
  table_name = "epi_units",
  eu_name = "NAME_2",
  geometry = "geom"
) |>
  extract_dataset()

entry_points_fp <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "entry_points",
    "BORDER_CROSSING_POINTS.csv"
  )

entry_points <- readr::read_csv(entry_points_fp)

entry_points <- validate_dataset(
  x = entry_points,
  table_name = "entry_points",
  point_name = "NAME",
  lng = "LONGITUDE_X",
  lat = "LATITUDE_Y",
  mode = "MODE",
  type = "TYPE",
  sources = "SOURCES"
) |>
  extract_dataset()

ri_entry_points <- calc_entry_point_risk(
  entry_points = entry_points,
  epi_units = tunisia,
  emission_risk = calc_emission_risk(
    emission_risk_factors = riskintrodata::get_wahis_erf(
      disease = "Anthrax",
      species = "Cattle",
      animal_category = "Domestic"
    )
  )
) |>
  rescale_risk_scores()

plot_risk(ri_entry_points, interactive = FALSE)
plot_risk(ri_entry_points, interactive = TRUE)
}
