% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-table.R
\name{remove_risk}
\alias{remove_risk}
\title{Remove risk from risk table}
\usage{
remove_risk(risk_table, cols)
}
\arguments{
\item{risk_table}{risk table to remove risk from}

\item{cols}{refers to columns (risk scores) in the risk table to remove}
}
\value{
\code{risk_table} with fewer associated risk scores
}
\description{
Remove risk from risk table
}
\seealso{
Other risk-table: 
\code{\link{add_risk}()},
\code{\link{risk_table}()}
}
\concept{risk-table}
