% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-table.R
\name{risk_table}
\alias{risk_table}
\title{Initialise table of risks}
\usage{
risk_table(epi_units, scale = c(0, 100))
}
\arguments{
\item{epi_units}{epidemiological units dataset}

\item{scale}{the scale to use for all risk scores added to the risk table. Each
new risk added to the table will be checked to ensure the same scale is used. For
example, a risk table using a scale from 0 to 100 will not accept a risk on a
scale from -1 to 1. This ensures comparability across risks.}
}
\value{
\code{sf} table with columns \code{eu_id}, \code{eu_name} and \code{geometry}, these all
come directly from the \code{epi_units} argument.
}
\description{
Creates the base table of risks where each risk associated with a study can be
stored. Each risk is checked that its on the same scale.
}
\seealso{
Other risk-table: 
\code{\link{add_risk}()},
\code{\link{remove_risk}()}
}
\concept{risk-table}
