% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rescaling.R
\name{inv_sigmoid}
\alias{inv_sigmoid}
\title{Inverse sigmoid function}
\usage{
inv_sigmoid(x)
}
\arguments{
\item{x}{Numeric vector with values between -1 and 1.}
}
\value{
A numeric vector with real values.
}
\description{
The invesrse of the sigmoid function.
}
\examples{
x <- c(0.01, 0.88, 0.247)
inv_sigmoid(x)
}
